refactore: restructure the project
